@@ -13,6 +13,7 @@
     "type-check": "tsc",
     "type-check:watch": "tsc --watch",
     "lint": "eslint .",
+    "lint:fix": "eslint --fix .",
     "test": "vitest",
     "test:once": "vitest run",
     "generate": "orval --config ./orval.config.ts"
@@ -20,52 +21,53 @@
   "dependencies": {
     "@badrap/result": "^0.2.13",
     "clsx": "^2.0.0",
-    "d3-scale": "^4.0.2",
-    "d3-scale-chromatic": "^3.0.0",
+    "d3": "^7.8.5",
     "date-fns": "^2.30.0",
     "fuse.js": "^6.6.2",
-    "immer": "^10.0.2",
+    "immer": "^10.0.3",
     "inter-ui": "^3.19.3",
     "little-state-machine": "^4.8.0",
     "lodash.debounce": "^4.0.8",
     "match-sorter": "^6.3.1",
+    "mousetrap": "^1.6.5",
     "react": "^18.2.0",
     "react-dnd": "^16.0.1",
     "react-dnd-html5-backend": "^16.0.1",
     "react-dom": "^18.2.0",
     "react-draggable": "^4.4.6",
-    "react-hook-form": "^7.46.2",
-    "react-hotkeys-hook": "^4.4.1",
+    "react-hook-form": "^7.47.0",
     "react-icons": "^4.11.0",
-    "react-router-dom": "^6.16.0",
+    "react-router-dom": "^6.17.0",
     "tiny-invariant": "^1.3.1",
     "use-immer": "^0.9.0",
-    "use-undo": "^1.1.1"
+    "use-undo": "^1.1.1",
+    "zod": "^3.22.4"
   },
   "devDependencies": {
-    "@testing-library/jest-dom": "^6.1.3",
+    "@testing-library/jest-dom": "^6.1.4",
     "@testing-library/react": "^14.0.0",
     "@testing-library/user-event": "^14.5.1",
-    "@types/d3-scale": "^4.0.5",
-    "@types/d3-scale-chromatic": "^3.0.0",
-    "@types/lodash.debounce": "^4.0.7",
-    "@types/react": "^18.2.23",
-    "@types/react-dom": "^18.2.8",
-    "@typescript-eslint/eslint-plugin": "^6.7.3",
-    "@typescript-eslint/parser": "^6.7.3",
+    "@types/d3": "^7.4.2",
+    "@types/lodash.debounce": "^4.0.8",
+    "@types/mousetrap": "^1.6.13",
+    "@types/react": "^18.2.31",
+    "@types/react-dom": "^18.2.14",
+    "@typescript-eslint/eslint-plugin": "^6.8.0",
+    "@typescript-eslint/parser": "^6.8.0",
     "@vitejs/plugin-react-swc": "^3.4.0",
     "autoprefixer": "^10.4.16",
-    "daisyui": "^3.8.1",
-    "eslint": "^8.50.0",
+    "daisyui": "^3.9.3",
+    "eslint": "^8.52.0",
     "eslint-plugin-react": "^7.33.2",
     "eslint-plugin-react-hooks": "^4.6.0",
     "jsdom": "^22.1.0",
-    "orval": "^6.17.0",
-    "postcss": "^8.4.30",
+    "orval": "^6.19.0",
+    "postcss": "^8.4.31",
     "prettier": "^3.0.3",
     "tailwindcss": "^3.3.3",
     "typescript": "^5.2.2",
-    "vite": "^4.4.9",
-    "vitest": "^0.34.5"
+    "vite": "^4.5.0",
+    "vite-tsconfig-paths": "^4.2.1",
+    "vitest": "^0.34.6"
   }
 }
