@@ -77,6 +77,7 @@ fn markers_to_clips(markers: Vec<Marker>) -> Vec<Clip> {
             range: (marker.start_time, marker.end_time),
             index_within_marker: 0,
             index_within_video: marker.index_within_video,
+            marker_title: marker.title.clone(),
         })
         .collect()
 }
@@ -104,6 +105,26 @@ pub struct ClipService {}
 impl ClipService {
     pub fn new() -> Self {
         Self {}
+    }
+
+    fn concatenate_clips(&self, clips: Vec<Clip>) -> Vec<Clip> {
+        let mut output = Vec::new();
+        let mut iter = clips.into_iter();
+
+        if let Some(mut current) = iter.next() {
+            for next in iter {
+                if current.range.1 == next.range.0 && current.video_id == next.video_id {
+                    current.range = (current.range.0, next.range.1);
+                } else {
+                    output.push(current);
+                    current = next;
+                }
+            }
+
+            output.push(current);
+        }
+
+        output
     }
 
     pub fn arrange_clips(&self, mut options: CreateClipsOptions) -> ClipsResult {
@@ -145,8 +166,19 @@ impl ClipService {
                 let sorter = RandomClipSorter;
                 sorter.sort_clips(clips, &mut rng)
             }
-            ClipOrder::SceneOrder => {
+            ClipOrder::Scene => {
                 let sorter = SceneOrderClipSorter;
+                sorter.sort_clips(clips, &mut rng)
+            }
+            ClipOrder::Fixed {
+                marker_title_groups,
+            } => {
+                let sorter = sort::FixedOrderClipSorter {
+                    marker_title_groups: marker_title_groups
+                        .into_iter()
+                        .map(|m| m.markers.into_iter().map(|s| s.title).collect())
+                        .collect(),
+                };
                 sorter.sort_clips(clips, &mut rng)
             }
             ClipOrder::NoOp => clips,
@@ -155,9 +187,22 @@ impl ClipService {
         let elapsed = start.elapsed();
         info!("generated {} clips in {:?}", clips.len(), elapsed);
 
+        let clips = self.concatenate_clips(clips);
+
         ClipsResult {
             clips,
             beat_offsets,
+        }
+    }
+}
+
+fn trim_clips(clips: &mut Vec<Clip>, max_len: f64) {
+    let clips_duration: f64 = clips.iter().map(|c| c.duration()).sum();
+    if clips_duration > max_len {
+        let slack = (clips_duration - max_len) / clips.len() as f64;
+        info!("clip duration {clips_duration} longer than permitted maximum duration {max_len}, making each clip {slack} shorter");
+        for clip in clips {
+            clip.range.1 -= slack;
         }
     }
 }
@@ -170,7 +215,7 @@ mod tests {
     use super::{ClipOrder, CreateClipsOptions};
     use crate::data::database::VideoSource;
     use crate::server::types::{
-        Clip, ClipOptions, ClipPickerOptions, EqualLengthClipOptions, PmvClipOptions,
+        Clip, ClipLengthOptions, ClipOptions, ClipPickerOptions, EqualLengthClipOptions,
         RandomizedClipOptions, RoundRobinClipOptions,
     };
     use crate::service::clip::sort::ClipSorter;
@@ -191,8 +236,9 @@ mod tests {
                 clip_picker: ClipPickerOptions::EqualLength(EqualLengthClipOptions {
                     clip_duration: 30.0,
                     divisors: vec![2.0, 3.0, 4.0],
+                    length: None,
                 }),
-                order: ClipOrder::SceneOrder,
+                order: ClipOrder::Scene,
             },
         };
         let service = ClipService::new();
@@ -214,7 +260,7 @@ mod tests {
             seed: None,
             clip_options: ClipOptions {
                 clip_picker: ClipPickerOptions::NoSplit,
-                order: ClipOrder::SceneOrder,
+                order: ClipOrder::Scene,
             },
         };
         let service = ClipService::new();
@@ -240,8 +286,9 @@ mod tests {
                 clip_picker: ClipPickerOptions::EqualLength(EqualLengthClipOptions {
                     clip_duration: 30.0,
                     divisors: vec![2.0, 3.0, 4.0],
+                    length: None,
                 }),
-                order: ClipOrder::SceneOrder,
+                order: ClipOrder::Scene,
             },
         };
 
@@ -274,6 +321,7 @@ mod tests {
                 range: (0.0, 9.0),
                 source: VideoSource::Folder,
                 video_id: "video".into(),
+                marker_title: "One".into(),
             },
             Clip {
                 index_within_marker: 0,
@@ -282,6 +330,7 @@ mod tests {
                 range: (1.0, 12.0),
                 source: VideoSource::Folder,
                 video_id: "video".into(),
+                marker_title: "Two".into(),
             },
         ];
         let mut rng = create_seeded_rng(None);
@@ -303,13 +352,14 @@ mod tests {
             seed: None,
             clip_options: ClipOptions {
                 clip_picker: ClipPickerOptions::RoundRobin(RoundRobinClipOptions {
-                    clip_lengths: PmvClipOptions::Randomized(RandomizedClipOptions {
+                    clip_lengths: ClipLengthOptions::Randomized(RandomizedClipOptions {
                         base_duration: 10.0,
                         divisors: vec![2.0, 3.0, 4.0],
                     }),
                     length: 30.0,
+                    lenient_duration: false,
                 }),
-                order: ClipOrder::SceneOrder,
+                order: ClipOrder::Scene,
             },
         };
         let service = ClipService::new();
@@ -329,13 +379,14 @@ mod tests {
             seed: None,
             clip_options: ClipOptions {
                 clip_picker: ClipPickerOptions::RoundRobin(RoundRobinClipOptions {
-                    clip_lengths: PmvClipOptions::Randomized(RandomizedClipOptions {
+                    clip_lengths: ClipLengthOptions::Randomized(RandomizedClipOptions {
                         base_duration: 10.0,
                         divisors: vec![2.0, 3.0, 4.0],
                     }),
                     length: 30.0,
+                    lenient_duration: false,
                 }),
-                order: ClipOrder::SceneOrder,
+                order: ClipOrder::Scene,
             },
         };
         let options = options.apply_marker_loops();
@@ -357,5 +408,53 @@ mod tests {
         let expected_length = 1084.0275;
         let total_duration: f64 = result.iter().map(|c| c.duration()).sum();
         assert_approx_eq!(f64, expected_length, total_duration, epsilon = 0.01);
+    }
+
+    #[test]
+    #[traced_test]
+    fn test_concatenate_clips() {
+        let clips = vec![
+            Clip {
+                index_within_marker: 0,
+                index_within_video: 0,
+                marker_id: 1,
+                range: (0.0, 9.0),
+                source: VideoSource::Folder,
+                video_id: "video".into(),
+                marker_title: "One".into(),
+            },
+            Clip {
+                index_within_marker: 0,
+                index_within_video: 1,
+                marker_id: 2,
+                range: (9.0, 12.0),
+                source: VideoSource::Folder,
+                video_id: "video".into(),
+                marker_title: "Two".into(),
+            },
+            Clip {
+                index_within_marker: 0,
+                index_within_video: 2,
+                marker_id: 3,
+                range: (12.0, 15.0),
+                source: VideoSource::Folder,
+                video_id: "video".into(),
+                marker_title: "Three".into(),
+            },
+            Clip {
+                index_within_marker: 0,
+                index_within_video: 3,
+                marker_id: 4,
+                range: (15.0, 18.0),
+                source: VideoSource::Folder,
+                video_id: "video2".into(),
+                marker_title: "Four".into(),
+            },
+        ];
+        let service = ClipService::new();
+        let concatenated = service.concatenate_clips(clips);
+        assert_eq!(concatenated.len(), 2);
+        assert_eq!(concatenated[0].range, (0.0, 15.0));
+        assert_eq!(concatenated[1].range, (15.0, 18.0));
     }
 }
