@@ -1,4 +1,4 @@
-use serde::Deserialize;
+use serde::{Deserialize, Serialize};
 use tracing::{debug, info};
 
 use super::ffmpeg::FfmpegLocation;
@@ -13,7 +13,7 @@ pub struct VideoParameters {
     pub codec: String,
 }
 
-#[derive(Default, Debug, Clone, PartialEq, Deserialize)]
+#[derive(Default, Debug, Clone, PartialEq, Deserialize, Serialize)]
 pub struct FfProbe {
     pub streams: Vec<Stream>,
     pub format: Format,
@@ -46,7 +46,7 @@ impl FfProbe {
     }
 }
 
-#[derive(Default, Debug, Clone, PartialEq, Deserialize)]
+#[derive(Default, Debug, Clone, PartialEq, Deserialize, Serialize)]
 pub struct Stream {
     pub index: i64,
     pub codec_name: Option<String>,
@@ -101,12 +101,12 @@ pub struct Stream {
     pub side_data_list: Vec<SideData>,
 }
 
-#[derive(Default, Debug, Clone, PartialEq, Deserialize)]
+#[derive(Default, Debug, Clone, PartialEq, Deserialize, Serialize)]
 pub struct SideData {
     pub side_data_type: String,
 }
 
-#[derive(Default, Debug, Clone, PartialEq, Deserialize)]
+#[derive(Default, Debug, Clone, PartialEq, Deserialize, Serialize)]
 pub struct Disposition {
     pub default: i64,
     pub dub: i64,
@@ -122,7 +122,7 @@ pub struct Disposition {
     pub timed_thumbnails: i64,
 }
 
-#[derive(Default, Debug, Clone, PartialEq, Deserialize)]
+#[derive(Default, Debug, Clone, PartialEq, Deserialize, Serialize)]
 pub struct StreamTags {
     pub language: Option<String>,
     pub creation_time: Option<String>,
@@ -130,7 +130,7 @@ pub struct StreamTags {
     pub encoder: Option<String>,
 }
 
-#[derive(Default, Debug, Clone, PartialEq, Deserialize)]
+#[derive(Default, Debug, Clone, PartialEq, Deserialize, Serialize)]
 pub struct Format {
     pub filename: String,
     pub nb_streams: i64,
@@ -153,7 +153,7 @@ impl Format {
     }
 }
 
-#[derive(Default, Debug, Clone, PartialEq, Deserialize)]
+#[derive(Default, Debug, Clone, PartialEq, Deserialize, Serialize)]
 pub struct FormatTags {
     #[serde(rename = "WMFSDKNeeded")]
     pub wmfsdkneeded: Option<String>,
@@ -170,10 +170,10 @@ pub struct FormatTags {
     pub encoder: Option<String>,
 }
 
-pub async fn ffprobe(path: &str, location: &FfmpegLocation) -> Result<FfProbe> {
+pub async fn ffprobe(path: impl AsRef<str>, location: &FfmpegLocation) -> Result<FfProbe> {
     use tokio::process::Command;
 
-    info!("running ffprobe on {path}");
+    info!("running ffprobe on {}", path.as_ref());
 
     let args = &[
         "-v",
@@ -182,7 +182,7 @@ pub async fn ffprobe(path: &str, location: &FfmpegLocation) -> Result<FfProbe> {
         "json",
         "-show_format",
         "-show_streams",
-        path,
+        path.as_ref(),
     ];
     debug!("running ffprobe with args {args:?}");
     let output = Command::new(location.ffprobe()).args(args).output().await?;
