@@ -1,40 +1,18 @@
-use std::collections::HashMap;
-
 use itertools::Itertools;
 use tracing::info;
 
-use super::commands::ffmpeg::FfmpegLocation;
-use super::commands::ffprobe::{ffprobe, FfProbe};
-use super::streams::{LocalVideoSource, StreamUrlService};
 use crate::data::database::Database;
 use crate::Result;
 
 #[derive(Clone)]
 pub struct EncodingOptimizationService {
-    ffmpeg_location: FfmpegLocation,
-    streams_service: StreamUrlService,
+    database: Database,
 }
 
 impl EncodingOptimizationService {
-    pub async fn new(ffmpeg_location: FfmpegLocation, database: Database) -> Self {
-        let streams_service = StreamUrlService::new(database).await;
-        EncodingOptimizationService {
-            ffmpeg_location,
-            streams_service,
-        }
-    }
-
-    async fn get_video_infos<'a>(
-        &self,
-        streams: &'a HashMap<String, String>,
-    ) -> HashMap<&'a str, Option<FfProbe>> {
-        let mut result = HashMap::new();
-        for (video_id, url) in streams {
-            let ffprobe = ffprobe(url, &self.ffmpeg_location).await;
-            result.insert(video_id.as_str(), ffprobe.ok());
-        }
-
-        result
+    pub fn new(database: Database) -> Self {
+        let service = Self { database };
+        service
     }
 
     pub async fn needs_re_encode(&self, video_ids: &[&str]) -> Result<bool> {
@@ -44,25 +22,15 @@ impl EncodingOptimizationService {
         }
 
         info!("checking encoding parameters for videos {:?}", video_ids);
+        let video_infos = self.database.ffprobe.get_infos(video_ids).await?;
 
-        let streams = self
-            .streams_service
-            .get_video_streams(video_ids, LocalVideoSource::File)
-            .await?;
-        let video_infos = self.get_video_infos(&streams).await;
-
-        let any_ffprobe_errors = video_infos.values().any(|v| v.is_none());
-        if any_ffprobe_errors {
-            Ok(true)
-        } else {
-            let parameters: Vec<_> = video_infos
-                .into_values()
-                .map(|v| v.unwrap().video_parameters())
-                .collect();
-            info!("video parameters: {:#?}", parameters);
-            let can_concatenate = parameters.into_iter().all_equal();
-            info!("can concatenate: {}", can_concatenate);
-            Ok(!can_concatenate)
-        }
+        let parameters: Vec<_> = video_infos
+            .into_iter()
+            .map(|v| v.video_parameters())
+            .collect();
+        info!("video parameters: {:#?}", parameters);
+        let can_concatenate = parameters.into_iter().all_equal();
+        info!("can concatenate: {}", can_concatenate);
+        Ok(!can_concatenate)
     }
 }
