@@ -1,8 +1,8 @@
 /**
- * Generated by orval v6.17.0 🍺
+ * Generated by orval v6.19.0 🍺
  * Do not edit manually.
  * clip-mash
- * OpenAPI spec version: 0.17.0
+ * OpenAPI spec version: 0.17.1
  */
 import {customInstance} from "./custom-client"
 export type GetHealthParams = {
@@ -44,6 +44,7 @@ export type ListVideosParams = {
   query?: string | null
   source?: VideoSource | null
   hasMarkers?: boolean | null
+  isInteractive?: boolean | null
   page?: number | null
   size?: number | null
   sort?: string | null
@@ -59,14 +60,24 @@ export type SplitMarkerParams = {
 
 export type DeleteMarker200 = unknown | null
 
+export type ListMarkerTitlesParams = {
+  count?: number | null
+  prefix?: string | null
+}
+
 export type ListMarkersParams = {
   videoIds?: string | null
+}
+
+export type ListFileEntriesParams = {
+  path?: string | null
+  includeHidden?: boolean | null
 }
 
 export type WeightedRandomClipOptionsWeightsItemItem = string & number
 
 export interface WeightedRandomClipOptions {
-  clipLengths: PmvClipOptions
+  clipLengths: ClipLengthOptions
   length: number
   weights: WeightedRandomClipOptionsWeightsItemItem[][]
 }
@@ -151,8 +162,6 @@ export type StrokeTypeOneOfThree = {
   accelerate: StrokeTypeOneOfThreeAccelerate
 }
 
-export type StrokeType = StrokeTypeOneOf | StrokeTypeOneOfThree
-
 /**
  * Creates a stroke every `n` beats
  */
@@ -164,6 +173,8 @@ export type StrokeTypeOneOf = {
   /** Creates a stroke every `n` beats */
   everyNth: StrokeTypeOneOfEveryNth
 }
+
+export type StrokeType = StrokeTypeOneOf | StrokeTypeOneOfThree
 
 export interface StashVideoDto {
   createdOn: number
@@ -213,12 +224,6 @@ export interface SongDto {
   url: string
 }
 
-export interface SongClipOptions {
-  beatsPerMeasure: number
-  cutAfterMeasures: MeasureCount
-  songs: Beats[]
-}
-
 export type SelectedMarkerSelectedRangeItem = number & number
 
 export interface SelectedMarker {
@@ -233,8 +238,9 @@ export interface SelectedMarker {
 }
 
 export interface RoundRobinClipOptions {
-  clipLengths: PmvClipOptions
+  clipLengths: ClipLengthOptions
   length: number
+  lenientDuration: boolean
 }
 
 export interface RandomizedClipOptions {
@@ -254,38 +260,6 @@ export interface Progress {
   message: string
   timestamp: string
   videoId: string
-}
-
-export type PmvClipOptionsOneOfFourAllOfType =
-  (typeof PmvClipOptionsOneOfFourAllOfType)[keyof typeof PmvClipOptionsOneOfFourAllOfType]
-
-// eslint-disable-next-line @typescript-eslint/no-redeclare
-export const PmvClipOptionsOneOfFourAllOfType = {
-  songs: "songs",
-} as const
-
-export type PmvClipOptionsOneOfFourAllOf = {
-  type: PmvClipOptionsOneOfFourAllOfType
-}
-
-export type PmvClipOptionsOneOfFour = SongClipOptions &
-  PmvClipOptionsOneOfFourAllOf
-
-export type PmvClipOptionsOneOf = RandomizedClipOptions &
-  PmvClipOptionsOneOfAllOf
-
-export type PmvClipOptions = PmvClipOptionsOneOf | PmvClipOptionsOneOfFour
-
-export type PmvClipOptionsOneOfAllOfType =
-  (typeof PmvClipOptionsOneOfAllOfType)[keyof typeof PmvClipOptionsOneOfAllOfType]
-
-// eslint-disable-next-line @typescript-eslint/no-redeclare
-export const PmvClipOptionsOneOfAllOfType = {
-  randomized: "randomized",
-} as const
-
-export type PmvClipOptionsOneOfAllOf = {
-  type: PmvClipOptionsOneOfAllOfType
 }
 
 export interface NewId {
@@ -308,6 +282,12 @@ export type MeasureCountOneOfThree = {
 
 export type MeasureCount = MeasureCountOneOf | MeasureCountOneOfThree
 
+export interface SongClipOptions {
+  beatsPerMeasure: number
+  cutAfterMeasures: MeasureCount
+  songs: Beats[]
+}
+
 export type MeasureCountOneOfType =
   (typeof MeasureCountOneOfType)[keyof typeof MeasureCountOneOfType]
 
@@ -319,6 +299,16 @@ export const MeasureCountOneOfType = {
 export type MeasureCountOneOf = {
   count: number
   type: MeasureCountOneOfType
+}
+
+export interface MarkerTitle {
+  count: number
+  title: string
+}
+
+export interface MarkerGroup {
+  markers: MarkerTitle[]
+  name: string
 }
 
 export interface MarkerDto {
@@ -346,6 +336,11 @@ export interface MarkerDtoPage {
   totalPages: number
 }
 
+export interface MarkerCount {
+  count: number
+  title: string
+}
+
 export interface ListVideoDto {
   markerCount: number
   video: VideoDto
@@ -359,9 +354,46 @@ export interface ListVideoDtoPage {
   totalPages: number
 }
 
+export interface ListFileEntriesResponse {
+  directory: string
+  entries: FileSystemEntry[]
+}
+
+export type FileSystemEntryOneOfThreeType =
+  (typeof FileSystemEntryOneOfThreeType)[keyof typeof FileSystemEntryOneOfThreeType]
+
+// eslint-disable-next-line @typescript-eslint/no-redeclare
+export const FileSystemEntryOneOfThreeType = {
+  file: "file",
+} as const
+
+export type FileSystemEntryOneOfThree = {
+  fileName: string
+  fullPath: string
+  size: number
+  type: FileSystemEntryOneOfThreeType
+}
+
+export type FileSystemEntryOneOfType =
+  (typeof FileSystemEntryOneOfType)[keyof typeof FileSystemEntryOneOfType]
+
+// eslint-disable-next-line @typescript-eslint/no-redeclare
+export const FileSystemEntryOneOfType = {
+  directory: "directory",
+} as const
+
+export type FileSystemEntryOneOf = {
+  fileName: string
+  fullPath: string
+  type: FileSystemEntryOneOfType
+}
+
+export type FileSystemEntry = FileSystemEntryOneOf | FileSystemEntryOneOfThree
+
 export interface EqualLengthClipOptions {
   clipDuration: number
   divisors: number[]
+  length?: number | null
 }
 
 export type EncodingEffort =
@@ -411,13 +443,6 @@ export interface CreateFunscriptBody {
   clips: Clip[]
 }
 
-export interface CreateClipsBody {
-  clipOrder: ClipOrder
-  clips: ClipOptions
-  markers: SelectedMarker[]
-  seed?: string | null
-}
-
 export interface CreateBeatFunscriptBody {
   songIds: number[]
   strokeType: StrokeType
@@ -443,12 +468,6 @@ export const ClipPickerOptionsOneOfOnezeroType = {
 export type ClipPickerOptionsOneOfOnezero = {
   type: ClipPickerOptionsOneOfOnezeroType
 }
-
-export type ClipPickerOptions =
-  | ClipPickerOptionsOneOf
-  | ClipPickerOptionsOneOfFour
-  | ClipPickerOptionsOneOfSeven
-  | ClipPickerOptionsOneOfOnezero
 
 export type ClipPickerOptionsOneOfSevenAllOfType =
   (typeof ClipPickerOptionsOneOfSevenAllOfType)[keyof typeof ClipPickerOptionsOneOfSevenAllOfType]
@@ -480,6 +499,12 @@ export type ClipPickerOptionsOneOfFourAllOf = {
 export type ClipPickerOptionsOneOfFour = WeightedRandomClipOptions &
   ClipPickerOptionsOneOfFourAllOf
 
+export type ClipPickerOptions =
+  | ClipPickerOptionsOneOf
+  | ClipPickerOptionsOneOfFour
+  | ClipPickerOptionsOneOfSeven
+  | ClipPickerOptionsOneOfOnezero
+
 export type ClipPickerOptionsOneOfAllOfType =
   (typeof ClipPickerOptionsOneOfAllOfType)[keyof typeof ClipPickerOptionsOneOfAllOfType]
 
@@ -495,19 +520,106 @@ export type ClipPickerOptionsOneOfAllOf = {
 export type ClipPickerOptionsOneOf = RoundRobinClipOptions &
   ClipPickerOptionsOneOfAllOf
 
-export type ClipOrder = (typeof ClipOrder)[keyof typeof ClipOrder]
+export type ClipOrderOneOfSevenType =
+  (typeof ClipOrderOneOfSevenType)[keyof typeof ClipOrderOneOfSevenType]
 
 // eslint-disable-next-line @typescript-eslint/no-redeclare
-export const ClipOrder = {
-  random: "random",
-  "scene-order": "scene-order",
+export const ClipOrderOneOfSevenType = {
+  fixed: "fixed",
+} as const
+
+export type ClipOrderOneOfSeven = {
+  markerTitleGroups: MarkerGroup[]
+  type: ClipOrderOneOfSevenType
+}
+
+export type ClipOrderOneOfFiveType =
+  (typeof ClipOrderOneOfFiveType)[keyof typeof ClipOrderOneOfFiveType]
+
+// eslint-disable-next-line @typescript-eslint/no-redeclare
+export const ClipOrderOneOfFiveType = {
   "no-op": "no-op",
 } as const
+
+export type ClipOrderOneOfFive = {
+  type: ClipOrderOneOfFiveType
+}
+
+export type ClipOrderOneOfThreeType =
+  (typeof ClipOrderOneOfThreeType)[keyof typeof ClipOrderOneOfThreeType]
+
+// eslint-disable-next-line @typescript-eslint/no-redeclare
+export const ClipOrderOneOfThreeType = {
+  scene: "scene",
+} as const
+
+export type ClipOrderOneOfThree = {
+  type: ClipOrderOneOfThreeType
+}
+
+export type ClipOrder =
+  | ClipOrderOneOf
+  | ClipOrderOneOfThree
+  | ClipOrderOneOfFive
+  | ClipOrderOneOfSeven
+
+export type ClipOrderOneOfType =
+  (typeof ClipOrderOneOfType)[keyof typeof ClipOrderOneOfType]
+
+// eslint-disable-next-line @typescript-eslint/no-redeclare
+export const ClipOrderOneOfType = {
+  random: "random",
+} as const
+
+export type ClipOrderOneOf = {
+  type: ClipOrderOneOfType
+}
 
 export interface ClipOptions {
   clipPicker: ClipPickerOptions
   order: ClipOrder
 }
+
+export interface CreateClipsBody {
+  clipOrder: ClipOrder
+  clips: ClipOptions
+  markers: SelectedMarker[]
+  seed?: string | null
+}
+
+export type ClipLengthOptionsOneOfFourAllOfType =
+  (typeof ClipLengthOptionsOneOfFourAllOfType)[keyof typeof ClipLengthOptionsOneOfFourAllOfType]
+
+// eslint-disable-next-line @typescript-eslint/no-redeclare
+export const ClipLengthOptionsOneOfFourAllOfType = {
+  songs: "songs",
+} as const
+
+export type ClipLengthOptionsOneOfFourAllOf = {
+  type: ClipLengthOptionsOneOfFourAllOfType
+}
+
+export type ClipLengthOptionsOneOfFour = SongClipOptions &
+  ClipLengthOptionsOneOfFourAllOf
+
+export type ClipLengthOptions =
+  | ClipLengthOptionsOneOf
+  | ClipLengthOptionsOneOfFour
+
+export type ClipLengthOptionsOneOfAllOfType =
+  (typeof ClipLengthOptionsOneOfAllOfType)[keyof typeof ClipLengthOptionsOneOfAllOfType]
+
+// eslint-disable-next-line @typescript-eslint/no-redeclare
+export const ClipLengthOptionsOneOfAllOfType = {
+  randomized: "randomized",
+} as const
+
+export type ClipLengthOptionsOneOfAllOf = {
+  type: ClipLengthOptionsOneOfAllOfType
+}
+
+export type ClipLengthOptionsOneOf = RandomizedClipOptions &
+  ClipLengthOptionsOneOfAllOf
 
 export type ClipRangeItem = number & number
 
@@ -515,6 +627,7 @@ export interface Clip {
   indexWithinMarker: number
   indexWithinVideo: number
   markerId: number
+  markerTitle: string
   /** Start and endpoint inside the video in seconds. */
   range: ClipRangeItem[]
   source: VideoSource
@@ -535,7 +648,7 @@ export const AddVideosRequestOneOfFiveType = {
 } as const
 
 export type AddVideosRequestOneOfFive = {
-  scene_ids: number[]
+  sceneIds: number[]
   type: AddVideosRequestOneOfFiveType
 }
 
@@ -571,6 +684,14 @@ export type AddVideosRequest =
   | AddVideosRequestOneOfThree
   | AddVideosRequestOneOfFive
 
+export const listFileEntries = (params?: ListFileEntriesParams) => {
+  return customInstance<ListFileEntriesResponse>({
+    url: `/api/library/directory`,
+    method: "get",
+    params,
+  })
+}
+
 export const listMarkers = (params?: ListMarkersParams) => {
   return customInstance<MarkerDto[]>({
     url: `/api/library/marker`,
@@ -585,6 +706,14 @@ export const createNewMarker = (createMarkerRequest: CreateMarkerRequest) => {
     method: "post",
     headers: {"Content-Type": "application/json"},
     data: createMarkerRequest,
+  })
+}
+
+export const listMarkerTitles = (params?: ListMarkerTitlesParams) => {
+  return customInstance<MarkerCount[]>({
+    url: `/api/library/marker/title`,
+    method: "get",
+    params,
   })
 }
 
@@ -808,15 +937,17 @@ export const getVersion = () => {
   return customInstance<Version>({url: `/api/version`, method: "get"})
 }
 
-type AwaitedInput<T> = PromiseLike<T> | T
-
-type Awaited<O> = O extends AwaitedInput<infer T> ? T : never
-
+export type ListFileEntriesResult = NonNullable<
+  Awaited<ReturnType<typeof listFileEntries>>
+>
 export type ListMarkersResult = NonNullable<
   Awaited<ReturnType<typeof listMarkers>>
 >
 export type CreateNewMarkerResult = NonNullable<
   Awaited<ReturnType<typeof createNewMarker>>
+>
+export type ListMarkerTitlesResult = NonNullable<
+  Awaited<ReturnType<typeof listMarkerTitles>>
 >
 export type UpdateMarkerResult = NonNullable<
   Awaited<ReturnType<typeof updateMarker>>
