@@ -1,11 +1,12 @@
 import {LoaderFunction} from "react-router-dom"
-import {getFormState} from "../helpers"
+import {getFormState} from "@/helpers"
 import {
   Clip,
   ClipPickerOptions,
   ClipsResponse,
   CreateClipsBody,
-  PmvClipOptions,
+  ListVideosParams,
+  ClipLengthOptions,
   StashVideoDtoPage,
   VideoDto,
   VideoSource,
@@ -14,73 +15,17 @@ import {
   listSongs,
   listStashVideos,
   listVideos,
-} from "../api"
-import {FormState} from "../types/form-state"
-import {getNewId, getVideo, listMarkers} from "../api"
+  getNewId,
+  getVideo,
+  listMarkers,
+} from "@/api"
+import {FormState} from "@/types/form-state"
+import {
+  ClipFormInputs,
+  getDefaultOptions,
+} from "./clips/settings/ClipSettingsForm"
 
 export const DEFAULT_PAGE_LENGTH = 24
-
-const getClipLengths = (state: FormState): PmvClipOptions => {
-  if (state.songs && state.songs.length) {
-    return {
-      type: "songs",
-      beatsPerMeasure: state.beatsPerMeasure || 4,
-      cutAfterMeasures: state.cutAfterMeasures || {type: "fixed", count: 4},
-      songs: state.songs.map((song) => ({
-        length: song.duration,
-        offsets: song.beats,
-      })),
-    }
-  } else {
-    return {
-      type: "randomized",
-      baseDuration: state.clipDuration || 30,
-      divisors: [2, 3, 4],
-    }
-  }
-}
-
-const getClipSettings = (state: FormState): ClipPickerOptions => {
-  if (state.clipStrategy === "weightedRandom") {
-    return {
-      type: "weightedRandom",
-      // @ts-expect-error form state needs to align with this
-      weights: state.clipWeights!,
-      clipLengths: getClipLengths(state),
-      length:
-        state.songs && state.songs.length > 0
-          ? state.songs.reduce((sum, song) => sum + song.duration, 0)
-          : state.selectedMarkers!.reduce(
-              (sum, {selectedRange: [start, end]}) => sum + (end - start),
-              0,
-            ),
-    }
-  } else if (state.clipStrategy === "equalLength") {
-    return {
-      type: "equalLength",
-      clipDuration: state.clipDuration || 30,
-      divisors: [2, 3, 4],
-    }
-  } else if (
-    state.clipStrategy === "roundRobin" &&
-    state.songs &&
-    state.songs.length > 0
-  ) {
-    return {
-      type: "roundRobin",
-      clipLengths: getClipLengths(state),
-      length: state.songs.reduce((sum, song) => sum + song.duration, 0),
-    }
-  } else if (state.clipStrategy === "noSplit" || state.splitClips === false) {
-    return {type: "noSplit"}
-  } else {
-    return {
-      type: "equalLength",
-      clipDuration: state.clipDuration || 30,
-      divisors: [2, 3, 4],
-    }
-  }
-}
 
 export interface ClipsLoaderData {
   clips: Clip[]
@@ -89,16 +34,118 @@ export interface ClipsLoaderData {
   beatOffsets?: number[]
 }
 
+const songsLength = (state: FormState): number => {
+  return state.songs?.reduce((len, song) => len + song.duration, 0) || 0
+}
+
+const markerLength = (state: FormState): number => {
+  return (
+    state.markers?.reduce(
+      (len, marker) => len + (marker.end - marker.start),
+      0,
+    ) || 0
+  )
+}
+
+const getClipLengths = (
+  options: {clipLengths?: ClipLengthOptions},
+  state: FormState,
+): ClipLengthOptions => {
+  if (!options.clipLengths || !options.clipLengths.type) {
+    return {
+      type: "randomized",
+      baseDuration: 20,
+      divisors: [2, 3, 4],
+    }
+  }
+
+  if (options.clipLengths.type === "songs") {
+    return {
+      type: "songs",
+      songs:
+        state.songs?.map((s) => ({
+          offsets: s.beats,
+          length: s.duration,
+        })) || [],
+      beatsPerMeasure: options.clipLengths.beatsPerMeasure,
+      cutAfterMeasures: options.clipLengths.cutAfterMeasures,
+    }
+  } else {
+    return {
+      type: "randomized",
+      baseDuration: options.clipLengths.baseDuration,
+      divisors: [2, 3, 4],
+    }
+  }
+}
+
+const getClipPickerOptions = (
+  inputs: ClipFormInputs | undefined,
+  state: FormState,
+): ClipPickerOptions => {
+  if (!inputs) {
+    return {
+      type: "equalLength",
+      clipDuration: 20,
+      divisors: [2, 3, 4],
+    }
+  }
+
+  switch (inputs.clipStrategy) {
+    case "roundRobin": {
+      const length = state.songs?.length
+        ? songsLength(state)
+        : markerLength(state)
+      return {
+        type: "roundRobin",
+        length,
+        clipLengths: getClipLengths(inputs.roundRobin, state),
+        lenientDuration: !inputs.useMusic,
+      }
+    }
+    case "weightedRandom": {
+      const length = state.songs?.length
+        ? songsLength(state)
+        : markerLength(state)
+      return {
+        type: "weightedRandom",
+        clipLengths: getClipLengths(inputs.weightedRandom, state),
+        length,
+        // @ts-expect-error type definitions don't align
+        weights: state.clipWeights!,
+      }
+    }
+    case "equalLength": {
+      const length = state.songs?.length ? songsLength(state) : undefined
+      return {
+        type: "equalLength",
+        clipDuration: inputs.equalLength.clipDuration,
+        divisors: [2, 3, 4],
+        length,
+      }
+    }
+    case "noSplit": {
+      return {
+        type: "noSplit",
+      }
+    }
+  }
+}
+
 export const clipsLoader: LoaderFunction = async () => {
   const state = getFormState()!
 
+  const options = state.clipOptions || getDefaultOptions(state)
+  const clipOrder = options.clipOrder || {type: "scene"}
+  const seed = options.seed?.length === 0 ? undefined : options.seed
+
   const body = {
-    clipOrder: state.clipOrder || "scene-order",
+    clipOrder,
     markers: state.selectedMarkers!.filter((m) => m.selected),
-    seed: state.seed || null,
+    seed,
     clips: {
-      clipPicker: getClipSettings(state),
-      order: state.clipOrder || "scene-order",
+      order: clipOrder,
+      clipPicker: getClipPickerOptions(options, state),
     },
   } satisfies CreateClipsBody
 
@@ -112,7 +159,7 @@ export const clipsLoader: LoaderFunction = async () => {
   return {
     ...data,
     videos,
-  } // satisfies ClipsLoaderData
+  }
 }
 
 export const localMarkerLoader: LoaderFunction = async () => {
@@ -161,18 +208,32 @@ export const stashVideoLoader: LoaderFunction = async ({request}) => {
   return videos
 }
 
-export const makeVideoLoader: (withMarkers: boolean) => LoaderFunction =
-  (withMarkers) =>
+function parseBoolean(str: string | null): boolean | null {
+  if (str === "true") {
+    return true
+  } else if (str === "false") {
+    return false
+  } else {
+    return null
+  }
+}
+
+export const makeVideoLoader: (
+  params: Partial<ListVideosParams>,
+) => LoaderFunction =
+  (params) =>
   async ({request}) => {
     const url = new URL(request.url)
     const query = url.searchParams
     const videos = await listVideos({
-      hasMarkers: withMarkers ? true : null,
+      hasMarkers: params.hasMarkers || parseBoolean(query.get("hasMarkers")),
       page: Number(query.get("page")) || 0,
       size: DEFAULT_PAGE_LENGTH,
       query: query.get("query"),
       sort: query.get("sort"),
       source: query.get("source") as VideoSource | null,
+      isInteractive:
+        params.isInteractive || parseBoolean(query.get("isInteractive")),
     })
 
     return videos
