@@ -1,15 +1,14 @@
+import {ClipFormInputs} from "@/routes/clips/settings/ClipSettingsForm"
 import {
   Clip,
   ClipOrder,
   EncodingEffort,
-  MeasureCount,
   SelectedMarker,
   VideoCodec,
   VideoQuality,
   SongDto,
   MarkerDto,
 } from "../api"
-import {ClipStrategy} from "./types"
 
 export enum FormStage {
   Start = 0,
@@ -22,30 +21,29 @@ export enum FormStage {
   CreateVideo = 7,
 }
 
+export type ClipOrderType = ClipOrder["type"]
+
 export interface FormState {
   stage: FormStage
   videoId?: string
   markers?: MarkerDto[]
   videoIds?: string[]
   recurse?: boolean
-  clipOrder?: ClipOrder
-  clipDuration?: number
   outputResolution?: [number, number]
   outputFps?: number
   selectedMarkers?: SelectedMarker[]
-  splitClips?: boolean
   fileName?: string
   clips?: Clip[]
   interactive?: boolean
-  seed?: string
   songs?: SongDto[]
   musicVolume?: number
-  trimVideoForSongs?: boolean
-  beatsPerMeasure?: number
-  cutAfterMeasures?: MeasureCount
-  clipStrategy?: ClipStrategy
-  clipWeights?: Array<[string, number]>
   videoCodec?: VideoCodec
   videoQuality?: VideoQuality
   encodingEffort?: EncodingEffort
+  clipWeights?: [string, number][]
+  clipOptions?: ClipFormInputs
+}
+
+export interface SerializedFormState extends FormState {
+  clipMashVersion: string
 }
