@@ -1,5 +1,6 @@
 import {defineConfig} from "vite"
 import react from "@vitejs/plugin-react-swc"
+import tsconfigPaths from "vite-tsconfig-paths";
 
 // https://vitejs.dev/config/
 export default defineConfig({
@@ -8,5 +9,5 @@ export default defineConfig({
       "/api": "http://localhost:5174",
     },
   },
-  plugins: [react()],
+  plugins: [react(), tsconfigPaths()],
 })
