@@ -1,6 +1,7 @@
 import * as path from "path"
 import {defineConfig} from "vitest/config"
 import react from "@vitejs/plugin-react-swc"
+import tsconfigPaths from "vite-tsconfig-paths";
 
 export default defineConfig({
   test: {
@@ -23,5 +24,5 @@ export default defineConfig({
       "~": path.resolve(__dirname, "app"),
     },
   },
-  plugins: [react()],
+  plugins: [react(), tsconfigPaths()],
 })
